+++ lib/rspec/support/warnings.rb
@@ -1,3 +1,6 @@
+require 'rspec/support'
+RSpec::Support.require_rspec_support "caller_filter"
+
 module RSpec
   module Support
     module Warnings