+++ Changelog.md
@@ -1,5 +1,8 @@
-### Development
-[Full Changelog](http://github.com/rspec/rspec-support/compare/v3.0.2...master)
+### 3.1.0 Development
+[Full Changelog](http://github.com/rspec/rspec-support/compare/v3.0.3...master)
+
+### 3.0.3 / 2014-07-21
+[Full Changelog](http://github.com/rspec/rspec-support/compare/v3.0.2...v3.0.3)
 
 Bug Fixes: